Rename rest-json-test packages
