+++ rest-json-testkit/src/test/scala/busymachines/rest_json_test/CRUDRoutesTest.scala
@@ -1,7 +1,7 @@
-package busymachines.rest_test
+package busymachines.rest_json_test
 
 import busymachines.rest._
-import busymachines.rest_test.routes._
+import busymachines.rest_json_test.routes_to_test._
 
 /**
   *
@@ -9,7 +9,7 @@ import busymachines.rest_test.routes._
   * @since 07 Sep 2017
   *
   */
-private[rest_test] class CRUDRoutesTest extends ExampleRestAPITestBaseClass {
+private[rest_json_test] class CRUDRoutesTest extends ExampleRestAPITestBaseClass {
   private lazy val crudAPI = new CRUDRoutesRestAPIForTesting()
   override implicit val testedRoute: Route = RestAPI.seal(crudAPI).route
   private implicit val cc: CallerContext = Contexts.none
